Rename module to 'transbank'
