+++ setup.cfg
@@ -2,7 +2,7 @@
 verbosity=1
 detailed-errors=1
 with-coverage=1
-cover-package=tbk
+cover-package=transbank
 
 [coverage:report]
 exclude_lines =
