--- tbk/__init__.py
@@ -1,4 +1,0 @@
-from tbk import *
-from tbk import onepay
-
-__all__ = ['onepay']
