+++ transbank/__init__.py
@@ -1,0 +1,4 @@
+from transbank import *
+from transbank import onepay
+
+__all__ = ['onepay']
